requirements.txt

pandas
requests
python-dotenv
psycopg2-binary
apache-airflow
google-ads
